Relocate `ThreadFormatter` to `Bosh::Director`
This consolidates the file to the (internal) gem that uses it. No usage
outside of `bosh` was found.
