+++ src/bosh-director/lib/bosh/director.rb
@@ -19,6 +19,8 @@ require 'time'
 require 'zlib'
 
 require 'common/exec'
+require 'common/core_ext/kernel'
+
 require 'bosh/template/evaluation_context'
 require 'bosh/version/release_version_list'
 
@@ -31,7 +33,6 @@ require 'nats/io/client'
 require 'openssl'
 require 'delayed_job_sequel'
 
-require 'common/thread_formatter'
 require 'bosh/director/cloud_factory'
 require 'bosh/director/az_cloud_factory.rb'
 require 'bosh/director/api'
@@ -56,6 +57,7 @@ require 'bosh/director/validation_helper'
 require 'bosh/director/download_helper'
 require 'bosh/director/formatter_helper'
 require 'bosh/director/tagged_logger'
+require 'bosh/director/thread_formatter'
 require 'bosh/director/duplicate_detector'
 require 'bosh/director/numerical_value_calculator'
 require 'bosh/director/regex_logging_filter'
