+++ src/bosh_common/lib/common/core_ext/kernel.rb
@@ -1,0 +1,9 @@
+module Kernel
+  def with_thread_name(name)
+    old_name = Thread.current[:name]
+    Thread.current[:name] = name
+    yield
+  ensure
+    Thread.current[:name] = old_name
+  end
+end
